refactor(structure): clean up structure & organize
This will help make Universal more approachable (since most people will
typically only be touching app.module etc. Moved some of the
services/workarounds, and separate the platform modules into their own
folder as well.

Tested dev / prod / ngc builds, all working.
