+++ src/platform-modules/node.module.ts
@@ -3,13 +3,13 @@ import { FormsModule } from '@angular/forms';
 import { RouterModule } from '@angular/router';
 import { UniversalModule, isBrowser, isNode } from 'angular2-universal/node'; // for AoT we need to manually split universal packages
 
-import { AppModule, AppComponent } from './+app/app.module';
-import { SharedModule } from './+app/shared/shared.module';
-import { CacheService } from './+app/shared/cache.service';
+import { AppModule, AppComponent } from '../+app/app.module';
+import { SharedModule } from '../+app/shared/shared.module';
+import { CacheService } from '../+app/shared/cache.service';
 
 // Will be merged into @angular/platform-browser in a later release
 // see https://github.com/angular/angular/pull/12322
-import { Meta } from './angular2-meta';
+import { Meta } from '../shared/meta/angular2-meta';
 
 export function getLRU() {
   return new Map();
