+++ src/shared/meta/angular2-meta.ts
