+++ src/shared/workarounds/__workaround.browser.ts
